Add preliminary support for the autocompletetype atttribute for phone/fax prefix and suffix.
BUG=92121
TEST=unit_tests --gtest_filter=FormStructureTest.*:AutofillManagerTest.*

Review URL: http://codereview.chromium.org/7694017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@97827 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7f512a0c118966c285254f037f68a952d6aeb0dc

+++ chrome/browser/autofill/phone_field.h
@@ -124,7 +124,7 @@ class PhoneField : public FormField {
   static struct Parser {
     RegexType regex;       // Field matching reg-ex.
     PhonePart phone_part;  // Index of the field.
-    int max_size;          // Max size of the field to match. 0 means any.
+    size_t max_size;       // Max size of the field to match. 0 means any.
   } phone_field_grammars_[];
 
   DISALLOW_COPY_AND_ASSIGN(PhoneField);
